+++ fabric/src/main/java/house/greenhouse/bovinesandbuttercups/content/attachment/BovinesAttachments.java
@@ -20,7 +20,7 @@ import java.util.UUID;
 public class BovinesAttachments {
     public static final AttachmentType<LockdownAttachment> LOCKDOWN = AttachmentRegistry.<LockdownAttachment>builder()
             .persistent(LockdownAttachment.CODEC)
-            .initializer(() -> new LockdownAttachment(Map.of()))
+            .initializer(() -> new LockdownAttachment(new HashMap<>()))
             .buildAndRegister(LockdownAttachment.ID);
     public static final AttachmentType<CowTypeAttachment> COW_TYPE = AttachmentRegistry.<CowTypeAttachment>builder()
             .persistent(CowTypeAttachment.CODEC)